macosx: compile fix for 10.4
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@23034 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/unix/macosx/cocoa/dialog/memorywindowcontroller.m
@@ -135,7 +135,7 @@
 
 - (id)tableView:(NSTableView *)aTableView
     objectValueForTableColumn:(NSTableColumn *)aTableColumn
-    row:(NSInteger)rowIndex
+    row:(int)rowIndex
 {
     if(data == nil)
         return nil;
@@ -230,7 +230,7 @@
     return theValue;
 }
 
-- (NSInteger)numberOfRowsInTableView:(NSTableView *)aTableView
+- (int)numberOfRowsInTableView:(NSTableView *)aTableView
 {
     return TOTAL_LINES;
 }
